Move a few classes around
