+++ libraries/apollo-runtime-java/src/main/java/com/apollographql/apollo3/runtime/java/ApolloClient.java
@@ -13,14 +13,17 @@ import com.apollographql.apollo3.api.Subscription;
 import com.apollographql.apollo3.api.http.DefaultHttpRequestComposer;
 import com.apollographql.apollo3.api.http.HttpHeader;
 import com.apollographql.apollo3.api.http.HttpMethod;
-import com.apollographql.apollo3.runtime.java.internal.AutoPersistedQueryInterceptor;
+import com.apollographql.apollo3.runtime.java.interceptor.ApolloInterceptor;
+import com.apollographql.apollo3.runtime.java.interceptor.ApolloInterceptorChain;
+import com.apollographql.apollo3.runtime.java.interceptor.internal.AutoPersistedQueryInterceptor;
+import com.apollographql.apollo3.runtime.java.interceptor.internal.DefaultInterceptorChain;
 import com.apollographql.apollo3.runtime.java.internal.DefaultApolloCall;
 import com.apollographql.apollo3.runtime.java.internal.DefaultApolloDisposable;
-import com.apollographql.apollo3.runtime.java.internal.DefaultInterceptorChain;
-import com.apollographql.apollo3.runtime.java.internal.http.HttpNetworkTransport;
-import com.apollographql.apollo3.runtime.java.internal.ws.GraphQLWsProtocol;
-import com.apollographql.apollo3.runtime.java.internal.ws.WebSocketNetworkTransport;
-import com.apollographql.apollo3.runtime.java.internal.ws.WsProtocol;
+import com.apollographql.apollo3.runtime.java.network.NetworkTransport;
+import com.apollographql.apollo3.runtime.java.network.http.HttpNetworkTransport;
+import com.apollographql.apollo3.runtime.java.network.ws.WebSocketNetworkTransport;
+import com.apollographql.apollo3.runtime.java.network.ws.protocol.GraphQLWsProtocol;
+import com.apollographql.apollo3.runtime.java.network.ws.protocol.WsProtocol;
 import okhttp3.Call;
 import okhttp3.OkHttpClient;
 import okhttp3.WebSocket;
