+++ libraries/apollo-runtime-java/src/main/java/com/apollographql/apollo3/runtime/java/interceptor/ApolloInterceptorChain.java
@@ -1,7 +1,9 @@
-package com.apollographql.apollo3.runtime.java;
+package com.apollographql.apollo3.runtime.java.interceptor;
 
 import com.apollographql.apollo3.api.ApolloRequest;
 import com.apollographql.apollo3.api.Operation;
+import com.apollographql.apollo3.runtime.java.ApolloCallback;
+import com.apollographql.apollo3.runtime.java.ApolloDisposable;
 import org.jetbrains.annotations.NotNull;
 
 /**
