+++ libraries/apollo-runtime-java/src/main/java/com/apollographql/apollo3/runtime/java/interceptor/internal/AutoPersistedQueryInterceptor.java
@@ -1,16 +1,16 @@
-package com.apollographql.apollo3.runtime.java.internal;
+package com.apollographql.apollo3.runtime.java.interceptor.internal;
 
 import com.apollographql.apollo3.api.ApolloRequest;
 import com.apollographql.apollo3.api.ApolloResponse;
+import com.apollographql.apollo3.api.Error;
 import com.apollographql.apollo3.api.Mutation;
 import com.apollographql.apollo3.api.Operation;
 import com.apollographql.apollo3.api.http.HttpMethod;
 import com.apollographql.apollo3.exception.ApolloException;
 import com.apollographql.apollo3.exception.AutoPersistedQueriesNotSupported;
 import com.apollographql.apollo3.runtime.java.ApolloCallback;
-import com.apollographql.apollo3.runtime.java.ApolloInterceptor;
-import com.apollographql.apollo3.runtime.java.ApolloInterceptorChain;
-import com.apollographql.apollo3.api.Error;
+import com.apollographql.apollo3.runtime.java.interceptor.ApolloInterceptor;
+import com.apollographql.apollo3.runtime.java.interceptor.ApolloInterceptorChain;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
