+++ libraries/apollo-runtime-java/src/main/java/com/apollographql/apollo3/runtime/java/network/ws/WebSocketConnection.java
@@ -1,4 +1,4 @@
-package com.apollographql.apollo3.runtime.java.internal.ws;
+package com.apollographql.apollo3.runtime.java.network.ws;
 
 import com.apollographql.apollo3.api.http.HttpHeader;
 import okhttp3.Headers;
