+++ libraries/apollo-runtime-java/src/main/java/com/apollographql/apollo3/runtime/java/network/ws/protocol/GraphQLWsProtocol.java
@@ -1,9 +1,10 @@
-package com.apollographql.apollo3.runtime.java.internal.ws;
+package com.apollographql.apollo3.runtime.java.network.ws.protocol;
 
 import com.apollographql.apollo3.api.ApolloRequest;
 import com.apollographql.apollo3.api.ImmutableMapBuilder;
 import com.apollographql.apollo3.api.Operation;
 import com.apollographql.apollo3.api.http.DefaultHttpRequestComposer;
+import com.apollographql.apollo3.runtime.java.network.ws.WebSocketConnection;
 
 import java.io.IOException;
 import java.util.Map;
@@ -45,7 +46,8 @@ public class GraphQLWsProtocol extends WsProtocol {
     this.pingIntervalMillis = pingIntervalMillis;
   }
 
-  @Override void connectionInit() {
+  @Override
+  public void connectionInit() {
     Map<String, Object> message = new ImmutableMapBuilder<String, Object>().put("type", "connection_init").build();
     Map<String, Object> payload = connectionPayload.get();
     if (payload != null) {
@@ -68,7 +70,7 @@ public class GraphQLWsProtocol extends WsProtocol {
 
   @SuppressWarnings("unchecked")
   @Override
-  void handleServerMessage(Map<String, Object> messageMap) {
+  public void handleServerMessage(Map<String, Object> messageMap) {
     String type = (String) messageMap.get("type");
     switch (type) {
       case "next":
@@ -96,7 +98,8 @@ public class GraphQLWsProtocol extends WsProtocol {
     }
   }
 
-  @Override <D extends Operation.Data> void startOperation(ApolloRequest<D> request) {
+  @Override
+  public <D extends Operation.Data> void startOperation(ApolloRequest<D> request) {
     sendMessageMap(
         new ImmutableMapBuilder<String, Object>()
             .put("type", "subscribe")
@@ -107,7 +110,8 @@ public class GraphQLWsProtocol extends WsProtocol {
     );
   }
 
-  @Override <D extends Operation.Data> void stopOperation(ApolloRequest<D> request) {
+  @Override
+  public <D extends Operation.Data> void stopOperation(ApolloRequest<D> request) {
     sendMessageMap(
         new ImmutableMapBuilder<String, Object>()
             .put("type", "complete")
@@ -117,7 +121,8 @@ public class GraphQLWsProtocol extends WsProtocol {
     );
   }
 
-  @Override protected void run() {
+  @Override
+  public void run() {
     if (pingIntervalMillis > 0) {
       pingExecutor.scheduleWithFixedDelay(() -> {
         Map<String, Object> map = new ImmutableMapBuilder<String, Object>()
