+++ tests/java-client/src/test/java/test/ClientTest.java
@@ -10,8 +10,8 @@ import com.apollographql.apollo3.mockserver.MockServer;
 import com.apollographql.apollo3.runtime.java.ApolloCallback;
 import com.apollographql.apollo3.runtime.java.ApolloClient;
 import com.apollographql.apollo3.runtime.java.ApolloDisposable;
-import com.apollographql.apollo3.runtime.java.ApolloInterceptor;
-import com.apollographql.apollo3.runtime.java.ApolloInterceptorChain;
+import com.apollographql.apollo3.runtime.java.interceptor.ApolloInterceptor;
+import com.apollographql.apollo3.runtime.java.interceptor.ApolloInterceptorChain;
 import com.google.common.truth.Truth;
 import io.reactivex.rxjava3.annotations.NonNull;
 import javatest.CreateCatMutation;
