+++ tests/java-client/src/test/java/test/GraphQLWsTest.java
@@ -5,9 +5,9 @@ import com.apollographql.apollo3.exception.ApolloException;
 import com.apollographql.apollo3.runtime.java.ApolloCallback;
 import com.apollographql.apollo3.runtime.java.ApolloClient;
 import com.apollographql.apollo3.runtime.java.ApolloDisposable;
-import com.apollographql.apollo3.runtime.java.internal.ws.GraphQLWsProtocol;
-import com.apollographql.apollo3.runtime.java.internal.ws.WebSocketNetworkTransport;
-import com.apollographql.apollo3.runtime.java.internal.ws.WsProtocol;
+import com.apollographql.apollo3.runtime.java.network.ws.WebSocketNetworkTransport;
+import com.apollographql.apollo3.runtime.java.network.ws.protocol.GraphQLWsProtocol;
+import com.apollographql.apollo3.runtime.java.network.ws.protocol.WsProtocol;
 import com.google.common.truth.Truth;
 import graphqlws.GreetingsSubscription;
 import graphqlws.HelloQuery;
